requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-extensions==3.1.1
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.7.3
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
zipp==3.4.1
